refactor: enhance presale program structure and error handling
- Refactored core program logic in lib.rs with improved instruction handling
- Enhanced configuration and state management in states/config.rs and states/position.rs
- Updated error handling with comprehensive error variants in error.rs
- Improved token purchase logic in instructions/buy_token.rs
- Enhanced token claim functionality in instructions/claim_token.rs
- Refined program initialization in instructions/initialize.rs
- Updated constants for better maintainability in constants.rs
- Updated dependencies in package.json and yarn.lock

+++ programs/fulbo-presale/src/constants.rs
@@ -13,7 +13,7 @@ pub const POSITION_SEED: &[u8] = b"position";
 
 /// One "month" for vesting purposes = 30 days in seconds.
 #[constant]
-pub const SECONDS_PER_MONTH: u64 = 30 * 24 * 60 * 60; // 2_592_000
+pub const SECONDS_PER_MONTH: u32 = 30 * 24 * 60 * 60; // 2_592_000
 
 /// Each month unlocks an additional 5% of total tokens (500 bps).
 #[constant]
